Green-a-thon
It's because of we provide a plate form where all type of people can(such as litterate and illitrate people) directly interact with ministory of  environment where they can know better knowlede and infirmation about our environment. They can also be able to share their personal solution and ask query and complain against environment issue to the ministory of environment.